Linux port

+++ lib/fecryptbuf.c
@@ -4,7 +4,7 @@ void fe_cryptbuf(char *buf, size_t bufsz, size_t offset, BitSequence *hashval) {
     size_t i;
 
     fe_msg(fectx(), "Transcrypting buffer of %lu bytes (offset %lu)\n",
-	bufsz, offset);
+	   (unsigned long)bufsz, (unsigned long)offset);
     
     for (i = 0; i < bufsz; i++) {
 	*buf ^= fe_randbyte(fectx(), offset, hashval);
